+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed loading stage order
 - Improved accuracy for loading timings log
 - Fixed loading progressbar appearance delay, i.e. no more visible progressbar when data instantly loaded
+- Fixed setting of `data` on `context` when a prepare handler returns a new value, before changes `context.data` contained a reference to an original (loaded) value, for now `Widget#context.data` is always equal to `Widget#data` after a `loadData()` call
 
 ## 1.0.0-beta.63 (07-02-2022)